Removing references to the rest of the lmfdb in the backend folder in preparation for extracting it to another module
