+++ lmfdb/backend/__init__.py
@@ -28,7 +28,7 @@ You can search using the methods ``search``, ``lucky`` and ``lookup``::
 
 
 from __future__ import absolute_import
-from .database import PostgresDatabase
+from .lmfdb_database import LMFDBDatabase
 from .utils import DelayCommit
 
 assert DelayCommit
@@ -36,4 +36,4 @@ from psycopg2.sql import SQL
 
 assert SQL
 
-db = PostgresDatabase()
+db = LMFDBDatabase()
