+++ lmfdb/backend/base.py
@@ -19,7 +19,6 @@ from psycopg2.sql import SQL, Identifier, Placeholder, Literal, Composable
 from psycopg2.extras import execute_values
 
 from .encoding import Json
-from lmfdb.logger import make_logger
 from .utils import reraise, DelayCommit, QueryLogFilter
 
 
@@ -214,16 +213,21 @@ class PostgresBase(object):
         # This function also sets self.conn
         db.register_object(self)
         self._db = db
-        from lmfdb.utils.config import Configuration
 
-        logging_options = Configuration().get_logging()
+        logging_options = db.config.options["logging"]
         self.slow_cutoff = logging_options["slowcutoff"]
-        handler = logging.FileHandler(logging_options["slowlogfile"])
+        self.logger = l = logging.getLogger(loggername)
+        l.propogate = False
+        l.setLevel(logging.INFO)
+        fhandler = logging.FileHandler(logging_options["slowlogfile"])
         formatter = logging.Formatter("%(asctime)s - %(message)s")
         filt = QueryLogFilter()
-        handler.setFormatter(formatter)
-        handler.addFilter(filt)
-        self.logger = make_logger(loggername, hl=False, extraHandlers=[handler])
+        fhandler.setFormatter(formatter)
+        fhandler.addFilter(filt)
+        l.addHandler(fhandler)
+        shandler = logging.StreamHandler()
+        shandler.setFormatter(formatter)
+        l.addHandler(shandler)
 
     def _execute(
             self,
