+++ lmfdb/utils/__init__.py
@@ -20,7 +20,6 @@ __all__ = ['request', 'make_response', 'flash', 'url_for', 'render_template',
            'debug', 'flash_error',
            'ajax_url',
            'image_callback', 'encode_plot',
-           'KeyedDefaultDict', 'make_tuple', 'range_formatter',
            'parse_ints', 'parse_signed_ints', 'parse_floats', 'parse_rational',
            'parse_rats', 'parse_bracketed_posints', 'parse_bracketed_rats', 'parse_bool',
            'parse_bool_unknown', 'parse_primes', 'parse_element_of',
@@ -30,7 +29,7 @@ __all__ = ['request', 'make_response', 'flash', 'url_for', 'render_template',
            'parse_galgrp', 'parse_nf_string', 'parse_subfield', 'parse_nf_elt',
            'parse_container', 'parse_hmf_weight', 'parse_count',
            'parse_start', 'parse_ints_to_list_flash', 'integer_options',
-           'nf_string_to_label', 'SearchParsingError', 'clean_input', 'prep_ranges',
+           'nf_string_to_label', 'clean_input', 'prep_ranges',
            'search_wrap', 'count_wrap',
            'SearchArray', 'TextBox', 'TextBoxNoEg', 'TextBoxWithSelect', 'BasicSpacer',
            'SkipBox', 'CheckBox', 'CheckboxSpacer', 'DoubleSelectBox', 'HiddenBox',
@@ -63,7 +62,6 @@ from .utilities import (
     debug, flash_error, 
     ajax_url,  # try to eliminate?
     image_callback, encode_plot,
-    KeyedDefaultDict, make_tuple, range_formatter,
     datetime_to_timestamp_in_ms, timestamp_in_ms_to_datetime)
 
 from .search_parsing import (
@@ -74,7 +72,7 @@ from .search_parsing import (
     parse_equality_constraints, parse_gap_id, parse_galgrp, parse_nf_string,
     parse_nf_elt, parse_container, parse_hmf_weight, parse_count, parse_start,
     parse_ints_to_list_flash, integer_options, nf_string_to_label, 
-    parse_subfield, SearchParsingError,
+    parse_subfield,
     clean_input, prep_ranges)
 
 from .search_wrapper import search_wrap, count_wrap
