+++ lmfdb/utils/display_stats.py
@@ -4,7 +4,8 @@ from collections import defaultdict
 from flask import url_for
 from sage.all import UniqueRepresentation, lazy_attribute, infinity
 
-from lmfdb.utils.utilities import format_percentage, display_knowl, KeyedDefaultDict, range_formatter
+from lmfdb.utils.utilities import format_percentage, display_knowl
+from lmfdb.backend.utils import KeyedDefaultDict, range_formatter
 
 class formatters(object):
     @classmethod
