+++ lmfdb/utils/utilities.py
@@ -11,7 +11,6 @@ import random
 import re
 import tempfile
 import time
-from collections import defaultdict
 from copy import copy
 from itertools import islice
 from types import GeneratorType
@@ -1204,56 +1203,6 @@ def encode_plot(P, pad=None, pad_inches=0.1, bbox_inches=None, remove_axes = Fal
     else:
         buf = virtual_file.buf
     return "data:image/png;base64," + quote(b64encode(buf))
-
-class KeyedDefaultDict(defaultdict):
-    """
-    A defaultdict where the default value takes the key as input.
-    """
-    def __missing__(self, key):
-        if self.default_factory is None:
-            raise KeyError((key,))
-        self[key] = value = self.default_factory(key)
-        return value
-
-def make_tuple(val):
-    """
-    Converts lists and dictionaries into tuples, recursively.  The main application
-    is so that the result can be used as a dictionary key.
-    """
-    if isinstance(val, (list, tuple)):
-        return tuple(make_tuple(x) for x in val)
-    elif isinstance(val, dict):
-        return tuple((make_tuple(a), make_tuple(b)) for a,b in val.items())
-    else:
-        return val
-
-def range_formatter(x):
-    if x is None:
-        return 'Unknown'
-    elif isinstance(x, dict):
-        if '$gte' in x:
-            a = x['$gte']
-        elif '$gt' in x:
-            a = x['$gt'] + 1
-        else:
-            a = None
-        if '$lte' in x:
-            b = x['$lte']
-        elif '$lt' in x:
-            b = x['$lt'] - 1
-        else:
-            b = None
-        if a == b:
-            return str(a)
-        elif b is None:
-            return "{0}-".format(a)
-        elif a is None:
-            return "..{0}".format(b)
-        else:
-            return "{0}-{1}".format(a,b)
-    return str(x)
-
-
 
 # conversion tools between timestamp different kinds of timestamp
 epoch = datetime.datetime.utcfromtimestamp(0)
